+++ src/modules/network_diagnostics/network_module.py
@@ -1046,9 +1046,9 @@ class NetworkDiagnosticsModule(CompositeModule):
         super().__init__()
         from modules.hosts_editor.hosts_editor_module import HostsEditorModule
         from modules.network_extras.net_extras_module import NetExtrasModule
-        from modules.wifi_analyzer.wifi_module import WifiAnalyzerModule
-        from modules.shared_resources.shares_module import SharesModule
         from modules.remote_tools.remote_module import RemoteToolsModule
+        from modules.shared_resources.shares_module import SharesModule
+        from modules.wifi_analyzer.wifi_module import WifiAnalyzerModule
 
         self.children = [
             NetworkToolsModule(),